requirements.txt
requests
beautifulsoup4
google-generativeai
gspread
google-auth
